url/requirements.txt
flask
PyJWT
hashids
python-dotenv
psycopg2-binary
sqlalchemy
requests